Task 14 is done

+++ src/part2/Task14.java
@@ -3,7 +3,7 @@ package part2;
 public class Task14 {
     public static void main(String[] args) {
         int i = 1;
-        int res = 1;
+        long res = 1;
         while (i<=25){
             res = res*i;
             i++;
